import bpsk library from `demod` branch
This gets BPSK working in realtime.

+++ src/bpsk/demod.h
@@ -13,7 +13,7 @@ struct bpsk_state;
 #define SAMPLES_PER_PERIOD 20 // Must evenly divide CARRIER_TONE
 
 void bpsk_demod_init(void);
-int bpsk_demod(int *bit, demod_sample_t *samples, size_t nb,
-               size_t *processed_samples);
+int bpsk_demod(int *bit, demod_sample_t *samples, uint32_t nb,
+               uint32_t *processed_samples);
 
 #endif
